separando botões do form checkout em componente e deixando checkout responsivo
