+++ src/components/forms/index.js
@@ -3,48 +3,15 @@ import { LocationForm } from "../Location";
 import Checkout from "../checkout";
 import { FormsContainer, HeaderForms, BodyForms } from "./style";
 import Cart from "../Cart";
-import { ButtonStyled } from "../../assets/style/buttonStyled";
-import { useContext } from "react";
-import { CartContext } from "../../contexts/CartContext";
-import useApi from "../../hooks/useApi";
-import useCart from "../../hooks/useCart";
+import { ButtonsForm } from "../ButtonsForm";
 
 const title = ["Carrinho", "Endereço", "Confirmar Compra"];
 
 export function FormsSteps() {
-  const { cart, address, orders } = useContext(CartContext);
-  const { city, street, district, number, paymentForms } = address;
-  const api = useApi(localStorage.getItem("token"));
-  const { deleteAllProducts } = useCart();
-
   function onSubmit(e) {
     e.preventDefault();
     next();
   }
-
-  async function postData() {
-    try {
-      const result = await api.checkoutPost(
-        city,
-        street,
-        district,
-        number,
-        paymentForms,
-        value,
-        orders
-      );
-      if (result.success) {
-        deleteAllProducts();
-        next();
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  const value = cart.reduce((total, { value }) => {
-    return (total += parseFloat(value));
-  }, 0);
 
   const { steps, currentStep, step, isFirstStep, back, next, isLastStep, goTo } =
     useMultStepFoms([<Cart />, <LocationForm />, <Checkout />, <Finished />]);
@@ -60,30 +27,14 @@ export function FormsSteps() {
       </HeaderForms>
       <BodyForms>
         {step}
-        {currentStep + 1 < 4 && (
-          <div>
-            {!isFirstStep && (
-              <ButtonStyled type="button" onClick={back}>
-                Voltar
-              </ButtonStyled>
-            )}
-            {cart.length ? (
-              <ButtonStyled
-                type={currentStep === 1 ? "submit" : "button"}
-                onClick={
-                  isLastStep ? postData : currentStep === 1 ? undefined : next
-                }
-              >
-                {isLastStep ? "Confirmar" : "Proximo"}
-              </ButtonStyled>
-            ) : (
-              ""
-            )}
-          </div>
-        )}
-        {currentStep === 3 ? <ButtonStyled onClick={() => {
-          goTo(0)
-        }}>Carrinho</ButtonStyled> : ""}
+        <ButtonsForm
+          currentStep={currentStep}
+          back={back}
+          isFirstStep={isFirstStep}
+          goTo={goTo}
+          isLastStep={isLastStep}
+          next={next}
+        />
       </BodyForms>
     </FormsContainer>
   );
